Add MultiExpSettings for multiexp_inner(), move into smaller modules.
